Renamed Anchor to AnchorProperties and added documentation.
